requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0
chromadb==1.3.4

# AWS
# Note: aioboto3 manages boto3/botocore versions, so we let it handle the dependencies
aioboto3==12.3.0

# AI/ML Services
google-generativeai==0.3.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# WebSocket
python-socketio==5.10.0
websockets==12.0

# Logging
python-json-logger==2.0.7

# HTTP Client
httpx>=0.27.0,<1.0.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Lambda Deployment
mangum==0.17.0

# Monitoring (Optional)
sentry-sdk==1.39.1
